# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S International Travel Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingItemDetails } from "./bookingItemDetails";
import type { BookingPaymentStatus } from "./bookingPaymentStatus";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: string;
  userId?: string;
  type: string;
  itemId: string;
  travelers: number;
  checkIn?: string;
  checkOut?: string;
  contactName: string;
  contactEmail: string;
  contactPhone: string;
  totalAmount: number;
  currency: string;
  status: BookingStatus;
  paymentStatus: BookingPaymentStatus;
  createdAt: string;
  itemDetails?: BookingItemDetails;
  notes?: string;
}