# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S International Travel Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingBodyType } from "./createBookingBodyType";

export interface CreateBookingBody {
  type: CreateBookingBodyType;
  itemId: string;
  travelers: number;
  checkIn?: string;
  checkOut?: string;
  contactName: string;
  contactEmail: string;
  contactPhone: string;
  totalAmount: number;
  currency: string;
  notes?: string;
}