# lib/api-zod/src/generated/types/holidayPackage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S International Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface HolidayPackage {
  id: string;
  title: string;
  destination: string;
  country?: string;
  theme: string;
  duration: number;
  price: number;
  currency: string;
  inclusions?: string[];
  highlights?: string[];
  imageUrl?: string;
  rating?: number;
  reviewCount?: number;
  maxTravelers?: number;
  description?: string;
  departureDates?: string[];
}