# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S International Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  phone?: string;
}